refact: update method returns

+++ src/main/java/tech/ada/school/managment/view/StudentController.java
@@ -23,17 +23,26 @@ public class StudentController {
     }
 
     @GetMapping
-    List<StudentDTO> getAll() {
-        return service.getAll();
+    public ResponseEntity<List<StudentDTO>> getAll() {
+        return ResponseEntity.ok().body(service.getAll());
     }
 
     @GetMapping("/{id}")
-    StudentDTO getById(@PathVariable("id") UUID id) {
-        return service.getById(id);
+    public ResponseEntity<StudentDTO> getById(@PathVariable("id") UUID id) {
+        try {
+            StudentDTO foundedStudent = service.getById(id);
+            if (foundedStudent != null) {
+                return ResponseEntity.ok().body(foundedStudent);
+            } else {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+            }
+        } catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
+        }
     }
 
     @PostMapping
-    ResponseEntity<StudentDTO> createStudent(@RequestBody @Valid StudentDTO student) {
+    public ResponseEntity<StudentDTO> createStudent(@RequestBody @Valid StudentDTO student) {
         StudentDTO createdStudent = service.createStudent(student);
         if (createdStudent != null) {
             return ResponseEntity.status(HttpStatus.CREATED).body(createdStudent);
@@ -43,17 +52,31 @@ public class StudentController {
     }
 
     @PutMapping("/{id}")
-    ResponseEntity<StudentDTO> updateStudent(@PathVariable("id") UUID id, @RequestBody @Valid StudentDTO student) {
-        StudentDTO updatedStudent = service.updateStudent(id, student);
-        if (updatedStudent != null) {
-            return ResponseEntity.status(HttpStatus.OK).body(updatedStudent);
-        } else {
+    public ResponseEntity<StudentDTO> updateStudent(@PathVariable("id") UUID id, @RequestBody @Valid StudentDTO student) {
+        try {
+            StudentDTO updatedStudent = service.updateStudent(id, student);
+            if (updatedStudent != null) {
+                return ResponseEntity.status(HttpStatus.OK).body(updatedStudent);
+            } else {
+                return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
+            }
+        } catch (Exception e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
         }
     }
 
     @DeleteMapping("/{id}")
-    void deleteStudent(@PathVariable("id") UUID id) {
-        service.deleteStudent(id);
+    public ResponseEntity deleteStudent(@PathVariable("id") UUID id) {
+        try {
+            StudentDTO foundedStudent = service.getById(id);
+            if (foundedStudent != null) {
+                service.deleteStudent(id);
+                return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+            } else {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+            }
+        } catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
+        }
     }
 }
